remove variable in phi
